linux-yocto-micro: Add 4.1 net-diet support
Add the remaining net-diet branch support for 4.1.  Note that part of
net-diet was added to the tinification branch to reduce conflicts,
this adds the rest, which it the bulk of it.

On the galileo gen2 board, With everything available from all branches
disabled, we still have functional TCP (ssh/scp, etc) networking and
the following size:

[ 0.000000] Memory: 222688K/245176K available (1567K kernel code, 131K
rwdata, 336K rodata, 244K init, 340K bss, 22488K reserved, 0K
cma-reserved)
[    0.000000] virtual kernel memory layout:
[    0.000000]     fixmap  : 0xfffa3000 - 0xfffff000   ( 368 kB)
[    0.000000]     vmalloc : 0xd05f0000 - 0xfffa1000   ( 761 MB)
[    0.000000]     lowmem  : 0xc0000000 - 0xcfdf0000   ( 253 MB)
[    0.000000]       .init : 0xc11ff000 - 0xc123c000   ( 244 kB)
[    0.000000]       .data : 0xc11881bc - 0xc11fde40   ( 471 kB)
[    0.000000]       .text : 0xc1000000 - 0xc11881bc   (1568 kB)

This doesn't include the several broken features in
KERNEL_FEATURES_BROKEN, which need to be fixed in a separate
iteration.

Signed-off-by: Tom Zanussi <tom.zanussi@linux.intel.com>